requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
future==0.18.3
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
serializers==0.2.4
sqlparse==0.4.4
tzdata==2023.3
